back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Proyecto Django para MVP Shopify Agent"
authors = [
    {name = "pablocevalenzuela",email = "pablocevalenzuela@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "langgraph (>=1.0.10,<2.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "supabase (>=2.28.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-environ (>=0.13.0,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
